requirements.txt
feedparser>=6.0.10
requests>=2.28.0
html2text>=2020.1.16
openai
pyyaml